% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_likelyhood_contour.R
\name{mixture_likelyhood_contour}
\alias{mixture_likelyhood_contour}
\title{Contour plot for likelyhood estimation of mixture distribution mutation rate estimation}
\usage{
mixture_likelyhood_contour(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist,
  trans_sites,
  resolution = 100,
  lambda_limits = c(1e-04, 1),
  k_limits = c(0.2, 0.99),
  estimate = NA,
  low_ci = NA,
  high_ci = NA,
  bins = NULL,
  title = "Mixture Likelihood Contour Plot"
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{resolution}{resolution of plot (main computation limitation)}

\item{lambda_limits}{bounds of lambda axis (plots on a log10 scale)}

\item{k_limits}{bounds of k axis}

\item{bins}{number of bins to pass to ggplot contour function}

\item{title}{title to pass to ggplot title}
}
\value{
ggplot contour map of likelyhood space
}
\description{
Contour plot for likelyhood estimation of mixture distribution mutation rate estimation
}
