% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_likelyhood_thresholds.R
\name{mixture_likelyhood_thresholds}
\alias{mixture_likelyhood_thresholds}
\title{Mixture Likelyhood Thresholds}
\usage{
mixture_likelyhood_thresholds(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist,
  trans_sites,
  truncation_point = 2000
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{year_range}{range of years to calculate thresholds over}
}
\value{
Threshold corresponding to the highest SNP distance where it is more likely to be related than unrelated
}
\description{
Mixture Likelyhood Thresholds
}
