% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_permutation_ci.R
\name{mixture_permutation_ci}
\alias{mixture_permutation_ci}
\title{Time Randomisation Test for mixture distribution mutation rate estimation}
\usage{
mixture_permutation_ci(
  trans_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  unrelated_snp_dist,
  unrelated_time_dist = NA,
  unrelated_sites = NA,
  percent_permute = 1,
  permutations = 3,
  sample_n = 500,
  confidence_level = 0.95,
  start_params = NA,
  ci_data = NA,
  title = NULL
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{unrelated_time_dist}{time differences for unrelated sample}

\item{unrelated_sites}{sites considered for unrelated sample}

\item{percent_permute}{proportion of mixed sample to permute}

\item{permutations}{number of permutation to run}

\item{sample_n}{number of bootstrap sampling to conduct}

\item{confidence_level}{confidence level for CIs}

\item{start_params}{initial parameters for optim, if NA (as default) will try 3 different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time.}

\item{ci_data}{optional input for previously calculated CI data (mixture_snp_cutoff_ci) for computational efficiency}

\item{title}{title for ggplot}
}
\value{
ggplot comparing point estimates and confidence levels between normal data and time randomised data
}
\description{
Time Randomisation Test for mixture distribution mutation rate estimation
}
