% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_snp_cutoff.R,
%   R/simulate_unrelsnp_data.R
\name{mixture_snp_cutoff}
\alias{mixture_snp_cutoff}
\title{Mixture SNP Cutoff}
\usage{
mixture_snp_cutoff(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  youden = FALSE,
  threshold_range = FALSE,
  max_time = NA,
  upper.tail = 0.95,
  max_false_positive = 0.05,
  trace = FALSE
)

mixture_snp_cutoff(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  youden = FALSE,
  threshold_range = FALSE,
  max_time = NA,
  upper.tail = 0.95,
  max_false_positive = 0.05,
  trace = FALSE
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{youden}{whether to produce SNP thresholds using the Youden method}

\item{threshold_range}{whether to produce a dataset of threshold considering a range of times (from 0.5 to 10 years)}

\item{max_time}{the maximum time(in days) utilised to calculate SNP thresholds, only applicable when time differences are provided}
}
\value{
SNP threshold, mutation rate, proportion related, estimated false positive and negative rate estimations

SNP threshold, mutation rate, proportion related, estimated false positive and negative rate estimations
}
\description{
Estimates evolutionary rates from a mixed transmission dataset and an unrelated data set with a mixed probability distribution approach.Uses these rates to produce a threshold of SNP distances to be considered linked or not.

Estimates evolutionary rates utilising a mixed transmission dataset and an unrelated data set with a mixed probability distribution approach.Uses these rates to produce a threshold of SNP distances to be considered linked or not.
}
