% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_snp_cutoff_ci.R
\name{mixture_snp_cutoff_ci}
\alias{mixture_snp_cutoff_ci}
\title{Bootstrapped confidence intervals for mixture SNP cutoffs}
\usage{
mixture_snp_cutoff_ci(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  sample_size = length(trans_snp_dist),
  sample_n = 1000,
  confidence_level = 0.95
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{sample_size}{size of each bootstrap sample}

\item{sample_n}{number of bootstrap sampling to conduct}
}
\value{
Confidence intervals
}
\description{
Creates confidence intervals for outputs from mixture_snp_cutoff using bootstrapping. Utilises the furrr package to allow for parallel computation, to enable use the plan() function from the future package.
}
