% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_snp_cutoff_NB.R
\name{mixture_snp_cutoff_NB}
\alias{mixture_snp_cutoff_NB}
\title{Mixture SNP Cutoff (Negative Binomial)}
\usage{
mixture_snp_cutoff_NB(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  youden = FALSE,
  threshold_range = FALSE,
  max_time = NA,
  upper.tail = 0.95,
  max_false_positive = 0.05,
  trace = FALSE,
  start_params = NA
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{youden}{whether to produce SNP thresholds using the Youden method}

\item{threshold_range}{whether to produce a dataset of threshold considering a range of times (from 0.5 to 10 years)}

\item{max_time}{the maximum time(in days) utilised to calculate SNP thresholds, only applicable when time differences are provided}

\item{upper.tail}{percentile to calculate SNP thresholds}

\item{max_false_positive}{if the false positive rate from calculated threshold is higher than this value a warning is produced}

\item{trace}{trace parameter to pass to optim}

\item{start_params}{initial parametrs for optim, if NA (as default) will try 3 different start parameters and produce the highest likelyhood result}
}
\value{
SNP threshold, mutation rate, proportion related, estimated false positive and negative rate estimations
}
\description{
Same as mixture_snp_cutoff but uses a negative binomial distribution to model the related SNP distances.
}
