% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_timerand_ci.R
\name{mixture_timerand_ci}
\alias{mixture_timerand_ci}
\title{Time Randomisation Test for mixture distribution mutation rate estimation}
\usage{
mixture_timerand_ci(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  truncation_point = NA,
  sample_size = length(trans_snp_dist),
  sample_n = 100,
  permutations = 3,
  quiet = FALSE,
  confidence_level = 0.95,
  within_individual = FALSE,
  subjectA_id,
  subjectB_id,
  clustered = FALSE,
  prop_type = "above_estimate",
  start_params = NA,
  ci_data = NA,
  title = NULL,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf)
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{truncation_point}{SNP distances to truncate at}

\item{sample_size}{size of each bootstrap sample}

\item{sample_n}{number of bootstrap sampling to conduct}

\item{permutations}{number of time permutation to run}

\item{quiet}{if false will print progress bar for each permutation}

\item{confidence_level}{confidence level for CIs}

\item{within_individual}{permute time data within each individuals (requires indiviudal ID codes that must be the exact same)}

\item{subjectA_id}{needed for within individual permutation}

\item{subjectB_id}{needed for within individual permutation (must be the same as subject A)}

\item{clustered}{permute between different time distances}

\item{prop_type}{proportion of bootstraps either "above_estimate", "above_low_ci", or "within_ci"}

\item{start_params}{initial parameters for optim, if NA (as default) will try 3 different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time.}

\item{ci_data}{optional input for previously calculated CI data (mixture_snp_cutoff_ci) for computational efficiency}

\item{title}{title for ggplot}
}
\value{
ggplot comparing point estimates and confidence levels between original data and time randomised data
}
\description{
Time Randomisation Test for mixture distribution mutation rate estimation
}
