% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_timerandboot.R
\name{mixture_timerandboot}
\alias{mixture_timerandboot}
\title{Mixture Timerandboot}
\usage{
mixture_timerandboot(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  ci_data = NA,
  title = NULL,
  sample_size = length(trans_snp_dist),
  truncation_point = NA,
  sample_n = 1000,
  confidence_level = 0.95,
  start_params = NA
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{sample_size}{size of each bootstrap sample}

\item{truncation_point}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit}

\item{sample_n}{number of bootstrap sampling to conduct}

\item{confidence_level}{confidence level to produce confidence intervals}

\item{start_params}{initial parametrs for optim, if NA (as default) will try 3 different start parameters and produce the highest likelyhood result. Specifying the start parameters minimises computing time.}
}
\value{
Confidence intervals
}
\description{
Creates confidence intervals for outputs from mixture_snp_cutoff using bootstrapping. Utilises the furrr package to allow for parallel computation, to enable use the plan() function from the future package.
}
