% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_estimate.R
\name{mxsure_estimate}
\alias{mxsure_estimate}
\title{Substituion rate estimation and SNP threshold inference}
\usage{
mxsure_estimate(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  right_truncation = 2000,
  threshold_time = NA,
  youden = FALSE,
  threshold_range = FALSE,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  return_tree_data = FALSE,
  branch_offset = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(0, Inf),
  tree_fulldist_param_bounds = c(0, Inf),
  upper.tail = 0.95,
  max_false_positive = 0.05,
  start_params = NA,
  trace = FALSE
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed related data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{right_truncation}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit. Will be faster with a lower truncation point.}

\item{threshold_time}{the time (in days) utilised to calculate SNP thresholds, only applicable when time differences are provided, if not provided will utilise maximum of supplied times}

\item{youden}{whether to produce additional SNP thresholds using the Youden method}

\item{threshold_range}{whether to produce a dataset of threshold considering a range of times (from 0.5 to 10 years)}

\item{tree}{SNP-scaled phylogenetic tree or list of trees with pairs of tips labelled with sampleA and sampleB. If this is supplied a different model for related SNP distances will be fit that takes into account branch length differences to the MRCA for any given pair of samples provided in sampleA and sampleB.}

\item{sampleA}{tip labels for sampleA; must be in the correct order with respect to sampleB such that the time distance is calculated as SampleA date-SampleB date (even if this allows for negative numbers)}

\item{sampleB}{tip labels for sampleB;see above}

\item{return_tree_data}{return data used for tree models instead of fitting model}

\item{branch_offset}{overide to branch offset to each skellam parameter for divergence correction model}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}

\item{tree_fulldist_param_bounds}{bounds of single branch fitting used in tree models}

\item{upper.tail}{percentile to calculate SNP thresholds}

\item{max_false_positive}{if the false positive rate from calculated threshold is higher than this value a warning is produced}

\item{start_params}{initial parameters for lambda, k, and intercept parameters for optimisation. If NA (as default) will try a range of different start parameters and produce the highest likelihood result}

\item{trace}{trace parameter to pass to nlminb}
}
\value{
Estimates for the related SNP threshold, substitution rate, proportion related, and estimated false positive rate
}
\description{
Estimates the substitution rate from SNP distances from a mixed related and an unrelated data set with a mixture distribution approach. Uses these rates to produce a threshold of SNP distances to be considered related or not.
}
\examples{
mixed_distances <- simulate_mixsnp_data(lambda=5, k=0.8, n=100)
distant_distances <- simulate_mixsnp_data(lambda=5, k=0, n=1000)
mxsure_estimate(mixed_snp_dist = mixed_distances$snp_dist,
unrelated_snp_dist = distant_distances$snp_dist,
mixed_time_dist = mixed_distances$time_dist)


}
