% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_input_instrain.R
\name{mxsure_input_instrain}
\alias{mxsure_input_instrain}
\title{Process inStrain compare input data}
\usage{
mxsure_input_instrain(
  input = NULL,
  file_path = NULL,
  dates = NULL,
  dates_sample_col = "sample_id",
  dates_date_col = "date"
)
}
\arguments{
\item{input}{A data.frame or data.table. If NULL, file_path is used.}

\item{file_path}{A string path to a file to be read by fread().}

\item{dates}{A data.frame or data.table with sample IDs and their collection dates.}

\item{dates_sample_col}{A string, the name of the sample ID column in 'dates'.}

\item{dates_date_col}{A string, the name of the date column in 'dates'.}
}
\value{
A data.table with columns: snp_dist, time_dist (if found or if sampling dates are provided), and sites (if found).
}
\description{
Efficiently reads or processes the output from the inStrain compare command to extract required columns
for MxSure It can calculate date differences between samples it by joining a provided 'dates'.
}
