% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_likelihood.R
\name{mxsure_likelihood}
\alias{mxsure_likelihood}
\title{mxsure likelihood estimation}
\usage{
mxsure_likelihood(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist,
  mixed_sites = NA,
  right_truncation = 2000,
  original_result = NA,
  start_params = NA,
  tree = NA,
  sampleA = NA,
  sampleB = NA
)
}
\arguments{
\item{mixed_snp_dist}{vector of SNP distances from mixed dataset}

\item{unrelated_snp_dist}{vector of SNP distances from unrelated dataset}

\item{mixed_time_dist}{vector of time differences for each SNP distacne in the mixed dataset}

\item{mixed_sites}{vector of sites considered for each SNP distance in the mixed dataset}

\item{right_truncation}{maximum limit of SNP distances to consider}

\item{original_result}{optional input to provide original mxsure_likelyhood() result instead of computing it again}

\item{start_params}{initial parameters for lambda, k, and intercept parameters for optimisation. If NA (as default) will try a range of different start parameters and produce the highest likelihood result}

\item{tree}{SNP-scaled phylogenetic tree or list of trees with pairs of tips labelled with sampleA and sampleB. If this is supplied a different model for related SNP distances will be fit that takes into account branch length differences to the MRCA for any given pair of samples provided in sampleA and sampleB.}

\item{sampleA}{tip labels for sampleA; must the earlier sampled branch}

\item{sampleB}{tip labels for sampleB; must be the later sample branch}
}
\value{
a dataframe with SNP distances, time differences, sites considered and the likeyhoods of related and unrelated models fitting for each datapoint
}
\description{
Reports likelihoods for related and unrelated distributions for each data point provided
}
\examples{
mixed_distances <- simulate_mixsnp_data(lambda=5, k=0.8, n=100)
distant_distances <- simulate_mixsnp_data(lambda=5, k=0, n=1000)
result <- mxsure_estimate(mixed_snp_dist = mixed_distances$snp_dist,
unrelated_snp_dist = distant_distances$snp_dist,
mixed_time_dist = mixed_distances$time_dist)

mxsure_likelihood(mixed_snp_dist = mixed_distances$snp_dist,
unrelated_snp_dist = distant_distances$snp_dist,
mixed_time_dist = mixed_distances$time_dist,
original_result = result)

}
