% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixsnp_data.R
\name{simulate_mixsnp_data}
\alias{simulate_mixsnp_data}
\title{Simulate Mixed related SNP distance dataset}
\usage{
simulate_mixsnp_data(
  lambda,
  k,
  unrel_shape = 100,
  unrel_rate = 1,
  error_param = NA,
  n = 100,
  rel_timelimit = 1
)
}
\arguments{
\item{lambda}{mutation rate input (SNPs per day)}

\item{k}{proportion of mixed dataset that is linked}

\item{unrel_shape}{shape of gamma distribution for time of evolution for unrelated pairs in years}

\item{unrel_rate}{rate of gamma distribution for time of evolution unrelated pairs in years}

\item{error_param}{chance that each SNP is missed}

\item{n}{size of dataset to simulate}

\item{rel_timelimit}{maximum time allowed between samples in years (calculates time difference from uniform dist.)}
}
\value{
SNP distance dataset with a mixture of related and unrelated pairs with time differences
}
\description{
Simulate Mixed related SNP distance dataset
}
