% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_snp_cutoff_ZK.R
\name{simulate_unrelsnp_data}
\alias{simulate_unrelsnp_data}
\title{Simulate unrelaed SNP distance dataset}
\usage{
simulate_unrelsnp_data(nbmu = 500, nbsize = 1, n = 1000)
}
\arguments{
\item{nbmu}{mean of the underlying negative binomial distribution}

\item{nbsize}{size parameter of the underlying negative binmial distribution (not the size of the simulated dataset)}

\item{n}{size/length of the simulated dataset}
}
\value{
unrelated SNP distance dataset
}
\description{
Simulate unrelaed SNP distance dataset
}
\examples{
y <- simulate_unrelsnp_data
}
