% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_snp_cutoff_ZK.R
\name{snp_hist}
\alias{snp_hist}
\title{SNP distace histogram}
\usage{
snp_hist(
  trans_snp_dist,
  unrelated_snp_dist = NULL,
  snp_threshold = NULL,
  limits = c(NA, 100)
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{snp_threshold}{a threshold to overlay on the graph (calculated with mixture_snp_cutoff)}

\item{limits}{snp distance/x axis limits}
}
\value{
a histogram plot of snp distances using ggplot
}
\description{
Creates a histogram of snp distances with or without the unrelated dataset
}
