% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_comparison_plot.R
\name{threshold_comparison_plot}
\alias{threshold_comparison_plot}
\title{Threshold Comparison Plot}
\usage{
threshold_comparison_plot(
  mixed_snp_dist,
  unrelated_snp_dist,
  snp_thresholds,
  method_names,
  show_related = FALSE,
  identity = FALSE,
  percent_dist_mixed,
  percent_dist_distant,
  identity_thresh = 99.99,
  k_est = NULL,
  title = NULL,
  labels = TRUE,
  dataset_names = c("Mixed", "Distant")
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{snp_thresholds}{list of SNP thresholds to consider}

\item{method_names}{list of names associated with each SNP threshold}

\item{show_related}{whether to show related dataset comparisons}

\item{identity}{whether to include an identity threshold as well}

\item{percent_dist_mixed}{percent distances from a mixed transmission data set, needed if identity thresholds are considered}

\item{percent_dist_distant}{percent distances from a distant transmission data set, needed if identity thresholds are considered}

\item{identity_thresh}{identity threshold to use (as percentage)}

\item{k_est}{proportion related line to plot}

\item{title}{title of plot}

\item{labels}{whether to include labels on the plot}

\item{dataset_names}{names of the mixed/distant datasets}
}
\value{
plot comparing proportion of data under each threshold
}
\description{
Threshold Comparison Plot
}
