% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncnbinom.R
\name{truncnbinom}
\alias{truncnbinom}
\alias{dtruncnbinom}
\alias{ptruncnbinom}
\alias{qtruncnbinom}
\title{Right truncated negative binomial distribution}
\usage{
dtruncnbinom(x, mu, size, truncation_point)

ptruncnbinom(q, mu, size, truncation_point)

qtruncnbinom(p, mu, size, truncation_point)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{mu}{mean of distribution}

\item{size}{target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{truncation_point}{right censoring point}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\description{
Density, distribution function, and quantile function for the right truncated negative binomial distribution with parameters size, prob, and truncation point.
}
