% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_snp_cutoff.R
\name{mixture_snp_cutoff}
\alias{mixture_snp_cutoff}
\title{Mixture SNP Cutoff}
\usage{
mixture_snp_cutoff(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  truncation_point = 2000,
  youden = FALSE,
  threshold_range = FALSE,
  max_time = NA,
  prior_lambda = NA,
  prior_k = NA,
  lambda_bounds = c(1e-04, 1),
  k_bounds = c(0, 1),
  upper.tail = 0.95,
  max_false_positive = 0.05,
  trace = FALSE,
  start_params = NA
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{truncation_point}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit. Will be faster with a lower truncation point.}

\item{youden}{whether to produce SNP thresholds using the Youden method}

\item{threshold_range}{whether to produce a dataset of threshold considering a range of times (from 0.5 to 10 years)}

\item{max_time}{the maximum time(in days) utilised to calculate SNP thresholds, only applicable when time differences are provided, if not provided will utilise maximum of supplied times}

\item{prior_lambda}{parameters for a gamma prior distribution for rate estimation, if set to "default" will use default parameters}

\item{prior_k}{parameters for a beta prior distribution for related proportion estimation, if set to "default" will use default parameters}

\item{lambda_bounds}{bounds of rate estimation}

\item{k_bounds}{bounds of related proportion estimation}

\item{upper.tail}{percentile to calculate SNP thresholds}

\item{max_false_positive}{if the false positive rate from calculated threshold is higher than this value a warning is produced}

\item{trace}{trace parameter to pass to nlminb}

\item{start_params}{initial parametrs for nlminb, if NA (as default) will try a range of different start parameters and produce the highest likelyhood result}
}
\value{
SNP threshold, mutation rate, proportion related, estimated false positive and negative rate estimations
}
\description{
Estimates evolutionary rates from a mixed transmission dataset and an unrelated data set with a mixed probability distribution approach.Uses these rates to produce a threshold of SNP distances to be considered linked or not.
}
