% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_timerand_ci.R
\name{mixture_timerand_ci}
\alias{mixture_timerand_ci}
\title{Time Randomisation Test for mixture distribution mutation rate estimation}
\usage{
mixture_timerand_ci(
  trans_snp_dist,
  unrelated_snp_dist,
  trans_time_dist = NA,
  trans_sites = NA,
  sample_size = length(trans_snp_dist),
  sample_n = 500,
  permutations = 3,
  confidence_level = 0.95,
  within_individual = FALSE,
  subjectA_id,
  subjectB_id,
  start_params = NA,
  ci_data = NA,
  raw = FALSE,
  title = NULL
)
}
\arguments{
\item{trans_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{trans_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{trans_sites}{list of sites considered for each SNP distance in mixed data set}

\item{sample_size}{size of each bootstrap sample}

\item{sample_n}{number of bootstrap sampling to conduct}

\item{confidence_level}{confidence level for CIs}

\item{start_params}{initial parameters for optim, if NA (as default) will try 3 different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time.}

\item{ci_data}{optional input for previously calculated CI data (mixture_snp_cutoff_ci) for computational efficiency}

\item{title}{title for ggplot}
}
\value{
ggplot comparing point estimates and confidence levels between normal data and time randomised data
}
\description{
Time Randomisation Test for mixture distribution mutation rate estimation
}
