% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_ci.R
\name{mxsure_ci}
\alias{mxsure_ci}
\title{Bootstrapped confidence intervals for related/unrelated strain SNP cutoffs}
\usage{
mxsure_ci(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  sample_size = length(mixed_snp_dist),
  right_truncation = NA,
  bootstraps = 100,
  confidence_level = 0.95,
  start_params = "Efficient",
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf),
  alpha_bounds = c(1e-10, Inf),
  beta_bounds = c(1e-10, Inf),
  quiet = FALSE
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{sample_size}{size of each bootstrap sample}

\item{right_truncation}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit}

\item{bootstraps}{number of bootstrap sampling to conduct}

\item{confidence_level}{confidence level to produce confidence intervals}

\item{start_params}{initial parameters for bootstrap estimates, if NA will try various different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time. If "Efficient" (as default) will use original result as starting paramters.}

\item{tree}{SNP-scaled phylogenetic tree or list of trees with pairs of tips labelled with sampleA and sampleB. If this is supplied a different model for related SNP distances will be fit that takes into account branch length differences to the MRCA for any given pair of samples provided in sampleA and sampleB.}

\item{sampleA}{tip labels for sampleA; must the earlier sampled branch}

\item{sampleB}{tip labels for sampleB; must be the later sample branch}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}

\item{alpha_bounds}{bounds for alpha paramter in phylo model}

\item{beta_bounds}{bounds for beta paramter in phylo model}

\item{quiet}{if true will not display progress bar}
}
\value{
Confidence intervals for all estimates produced by mxsure_estimate.
}
\description{
Creates confidence intervals for outputs from mxsure_estimate using bootstrapping. Utilises the furrr package to allow for parallel computation, to enable use the plan() function from the future package.
}
\examples{
mixed_distances <- simulate_mixsnp_data(lambda=5, k=0.8, n=100)
distant_distances <- simulate_mixsnp_data(lambda=5, k=0, n=1000)
result <- mxsure_estimate(mixed_snp_dist = mixed_distances$snp_dist,
unrelated_snp_dist = distant_distances$snp_dist,
mixed_time_dist = mixed_distances$time_dist)

mxsure_ci(mixed_snp_dist = mixed_distances$snp_dist,
unrelated_snp_dist = distant_distances$snp_dist,
mixed_time_dist = mixed_distances$time_dist,
bootstraps=5) #practical useage would require more bootstraps

}
