% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_ci.R
\name{mxsure_ci}
\alias{mxsure_ci}
\title{Bootstrapped confidence intervals for mixture SNP cutoffs}
\usage{
mxsure_ci(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  sample_size = length(mixed_snp_dist),
  truncation_point = NA,
  sample_n = 100,
  confidence_level = 0.95,
  start_params = "Efficient",
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf),
  shared_snp_lambda_bounds = c(0, Inf),
  shared_snp_intercept_bounds = c(0, Inf),
  quiet = FALSE
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{sample_size}{size of each bootstrap sample}

\item{truncation_point}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit}

\item{sample_n}{number of bootstrap sampling to conduct}

\item{confidence_level}{confidence level to produce confidence intervals}

\item{start_params}{initial parametrs for optim, if NA (as default) will try 3 different start parameters and produce the highest likelyhood result. Specifying the start parameters minimises computing time.}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}
}
\value{
Confidence intervals for all estimates produced by mxsure_estimate.
}
\description{
Creates confidence intervals for outputs from mxsure_estimate using bootstrapping. Utilises the furrr package to allow for parallel computation, to enable use the plan() function from the future package.
}
