% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_estimate.R
\name{mxsure_estimate}
\alias{mxsure_estimate}
\title{mxsure}
\usage{
mxsure_estimate(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  truncation_point = 2000,
  youden = FALSE,
  threshold_range = FALSE,
  max_time = NA,
  start_params = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf),
  prior_lambda = NA,
  prior_k = NA,
  upper.tail = 0.95,
  max_false_positive = 0.05,
  trace = FALSE
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed related data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{truncation_point}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit. Will be faster with a lower truncation point.}

\item{youden}{whether to produce additional SNP thresholds using the Youden method}

\item{threshold_range}{whether to produce a dataset of threshold considering a range of times (from 0.5 to 10 years)}

\item{max_time}{the time (in days) utilised to calculate SNP thresholds, only applicable when time differences are provided, if not provided will utilise maximum of supplied times}

\item{start_params}{initial parameters for lambda, k, and intercept parameters for optimisation. If NA (as default) will try a range of different start parameters and produce the highest likelyhood result}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}

\item{prior_lambda}{parameters for a gamma prior distribution for rate estimation, if set to "default" will use default parameters}

\item{prior_k}{parameters for a beta prior distribution for related proportion estimation, if set to "default" will use default parameters}

\item{upper.tail}{percentile to calculate SNP thresholds}

\item{max_false_positive}{if the false positive rate from calculated threshold is higher than this value a warning is produced}

\item{trace}{trace parameter to pass to nlminb}
}
\value{
Estimates for the related SNP threshold, substitution rate, proportion related, and estimated false positive rate
}
\description{
Estimates the substitution rate from SNP distances from a mixed related and an unrelated data set with a mixture distribution approach. Uses these rates to produce a threshold of SNP distances to be considered related or not.
}
\examples{
x <- simulate_mixsnp_data(1, 0.8)
y<- simulate_mixsnp_data(1, 0, n=1000)
mxsure_estimate(x$snp_dist, y$snp_dist, x$time_dist)


}
