% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_input_distmatrix.R
\name{mxsure_input_distmatrix}
\alias{mxsure_input_distmatrix}
\title{Process a TRACS distance matrix for MxSure}
\usage{
mxsure_input_distmatrix(
  input = NULL,
  file_path = NULL,
  dates = NULL,
  dates_sample_col = "sample_id",
  dates_date_col = "date"
)
}
\arguments{
\item{input}{A data.frame or data.table (distance matrix). If NULL, file_path is used.}

\item{file_path}{A string path to a file to be read by fread() (distance matrix).}

\item{dates}{A data.frame or data.table with sample IDs and their collection dates.}

\item{dates_sample_col}{A string, the name of the sample ID column in 'dates'.}

\item{dates_date_col}{A string, the name of the date column in 'dates'.}

\item{rownames}{If input has row names (TRUE) or if column one are sample names (FALSE)}
}
\value{
A data.table with columns: sampleA, sampleB, snp_dist,
and time_dist (if 'dates' was provided).
}
\description{
Efficiently reads a distance matrix, "melts" it into a long-form pairwise
table, and optionally joins a 'dates' table to calculate time differences.
}
