% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_input_tree.R
\name{mxsure_input_tree}
\alias{mxsure_input_tree}
\title{Process a Phylogenetic Tree for MxSure}
\usage{
mxsure_input_tree(
  tree = NULL,
  file_path = NULL,
  tip_dates = NULL,
  tip_cluster_ids = NULL
)
}
\arguments{
\item{tree}{A 'phylo' object (from the 'ape' package). If NULL, file_path is used.}

\item{file_path}{A string path to a tree file (e.g., Newick format) to be read by ape::read.tree().}

\item{tip_dates}{A vector of dates corresponding to the order of tips in the tree.}

\item{tip_cluster_ids}{A vector of related clusters that, if provided, will flag which comparisons are from the same cluster or not. This could be individual IDs for tagging within-host comparisons or specific transmission clusters or sites.}
}
\value{
A data.table with columns: sampleA, sampleB, snp_dist, time_dist (if 'tip_dates' was provided), and same_cluster (if 'tip_cluster_ids' provided).
}
\description{
Efficiently reads a phylogenetic tree, calculates pairwise tip distances,
checks for valid SNP-scale branch lengths, rounds them to integers,
and optionally maps a vector of dates to calculate time differences.
}
