% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_likelyhood.R
\name{mxsure_likelyhood}
\alias{mxsure_likelyhood}
\title{Reports Likelyhoods for related and unrelated distributions for each datapoint provided}
\usage{
mxsure_likelyhood(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist,
  mixed_sites,
  truncation_point = 2000,
  original_result = NA,
  start_params = NA,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  branch_offset = NA
)
}
\arguments{
\item{mixed_snp_dist}{vector of SNP distances from mixed dataset}

\item{unrelated_snp_dist}{vector of SNP distances from unrelated dataset}

\item{mixed_time_dist}{vector of time differences for each SNP distacne in the mixed dataset}

\item{mixed_sites}{vector of sites considered for each SNP distance in the mixed dataset}

\item{truncation_point}{maximum limit of SNP distances to consider}
}
\value{
a dataframe with SNP distances, time differences, sites considered and the likeyhoods of related and unrelated models fitting for each datapoint
}
\description{
Reports Likelyhoods for related and unrelated distributions for each datapoint provided
}
