% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_timerandtest.R
\name{mxsure_timerandtest}
\alias{mxsure_timerandtest}
\title{Time Randomisation Test for mixture distribution mutation rate estimation}
\usage{
mxsure_timerandtest(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  truncation_point = 2000,
  sample_size = length(mixed_snp_dist),
  sample_n = 100,
  permutations = 5,
  quiet = FALSE,
  confidence_level = 0.95,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  start_params = "Efficient",
  original_result = NA,
  ci_data = NA,
  title = NULL,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf),
  single_branch_lambda_bounds = c(0, Inf),
  branch_offset = NA,
  within_individual = FALSE,
  subjectA_id,
  subjectB_id,
  clustered = FALSE,
  prop_type = "above_estimate"
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{truncation_point}{SNP distances to truncate at}

\item{sample_size}{size of each bootstrap sample}

\item{sample_n}{number of bootstrap sampling to conduct for each time permutation}

\item{permutations}{number of time permutation to run}

\item{quiet}{if false will print progress bar for each permutation}

\item{confidence_level}{confidence level for CIs}

\item{start_params}{initial parameters for optim, if NA (as default) will try 3 different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time.}

\item{ci_data}{optional input for previously calculated CI data (mxsure_ci) for computational efficiency}

\item{title}{title for ggplot}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}

\item{within_individual}{permute time data within each individuals (requires indiviudal ID codes that must be the exact same)}

\item{subjectA_id}{needed for within individual permutation}

\item{subjectB_id}{needed for within individual permutation (must be the same as subject A)}

\item{clustered}{permute between different time distances}

\item{prop_type}{proportion of bootstraps either "above_estimate", "above_low_ci", or "within_ci"}
}
\value{
list of overall outcomes, results from each time randomisation, raw results, and a plot comparing point estimates and confidence levels between original data and time randomised data
}
\description{
Time Randomisation Test for mixture distribution mutation rate estimation
}
