% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsure_timerandtest.R
\name{mxsure_timerandtest}
\alias{mxsure_timerandtest}
\title{Time Randomisation Test for mixture distribution substituion rate estimation}
\usage{
mxsure_timerandtest(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist = NA,
  mixed_sites = NA,
  right_truncation = 2000,
  bootstraps = 100,
  permutations = 5,
  quiet = FALSE,
  confidence_level = 0.95,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  start_params = "Efficient",
  original_result = NA,
  ci_data = NA,
  title = NULL,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf),
  tree_fulldist_param_bounds = c(0, Inf),
  branch_offset = NA,
  failure_criterion = "above_estimate"
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{right_truncation}{SNP distances to truncate at}

\item{bootstraps}{number of bootstrap sampling to conduct for each time permutation}

\item{permutations}{number of time permutation to run}

\item{quiet}{if false will print progress bar for each permutation}

\item{confidence_level}{confidence level for CIs}

\item{tree}{SNP-scaled phylogenetic tree or list of trees with pairs of tips labelled with sampleA and sampleB. If this is supplied a different model for related SNP distances will be fit that takes into account branch length differences to the MRCA for any given pair of samples provided in sampleA and sampleB.}

\item{sampleA}{tip labels for sampleA; must be in the correct order with respect to sampleB such that the time distance is calculated as SampleA date-SampleB date (even if this allows for negative numbers)}

\item{sampleB}{tip labels for sampleB;see above}

\item{start_params}{initial parameters for bootstrap estimates, if NA will try various different start parameters and produce the highest likelihood result. Specifying the start parameters minimises computing time. If "Efficient" (as default) will use original result as starting paramters.}

\item{original_result}{if supplied the function will not re-estimate the original estimate}

\item{ci_data}{optional input for previously calculated CI data (mxsure_ci) for computational efficiency}

\item{title}{title for ggplot}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}

\item{tree_fulldist_param_bounds}{bounds of single branch fitting used in tree models}

\item{branch_offset}{overide to branch offset to each skellam parameter for divergence correction model}

\item{failure_criterion}{proportion of bootstraps either "above_estimate", "above_low_ci", or "within_ci"}
}
\value{
list of overall outcomes, results from each time randomisation, raw results, and a plot comparing point estimates and confidence levels between original data and time randomised data. Failure bootstraps and failure percentages indicate the number and percentage of bootstrap estimates from time randomised datasets above the original estimate (if prop type is not adjusted)
}
\description{
Time Randomisation Test for mixture distribution substituion rate estimation
}
