% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixsnp_data.R
\name{simulate_mixsnp_data}
\alias{simulate_mixsnp_data}
\title{Simulate Mixed related SNP distance dataset}
\usage{
simulate_mixsnp_data(lambda, k, nbmu = 500, error_param = NA, n = 1000)
}
\arguments{
\item{lambda}{mutation rate input (SNPs per day)}

\item{k}{proportion of mixed dataset that is linked}

\item{nbmu}{mean of the negative binomial distribution underlying the unrelated SNP distances}

\item{error_param}{chance that each SNP is missed}

\item{n}{size of dataset to simulate}
}
\value{
SNP distance dataset with a mixture of related and unrelated pairs with time differences
}
\description{
Simulate Mixed related SNP distance dataset
}
\examples{
#simulates dataset with an E. coli profile
x <- simulate_mixsnp_data(lambda=0.00179*4, 0.5, 500, error_param=NA)
}
