% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_over_time.R
\name{snp_over_time}
\alias{snp_over_time}
\title{SNP distance/time plot}
\usage{
snp_over_time(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist,
  mixed_sites = NA,
  right_truncation = 2000,
  original_result = NA,
  start_params = NA,
  threshold_time = NA,
  title = "SNP Distance Over Time",
  jitter = TRUE,
  failure_perc = NA,
  ci_data = NA,
  time_limits = c(0, NA),
  under_threshold = FALSE,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  branch_offset = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf)
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{right_truncation}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit}

\item{original_result}{optional input to provide original mxsure_likelyhood() result instead of computing it again}

\item{start_params}{initial parameters for lambda, k, and intercept parameters for optimisation. If NA (as default) will try a range of different start parameters and produce the highest likelihood result}

\item{threshold_time}{the time (in days) utilised to calculate SNP thresholds, only applicable when time differences are provided, if not provided will utilise maximum of supplied times}

\item{title}{title for graph passed to ggplot}

\item{jitter}{whether to jitter data (keeps data above 0 SNPs and 0 time)}

\item{failure_perc}{optional input to display time randomised failure percentage from mxsure_timerand_test}

\item{ci_data}{optional input for previously calculated CI data (mxsure_ci) to display on subtitle}

\item{time_limits}{x axis limits passed to ggplot}

\item{under_threshold}{whether to only plot points below the calculated SNP threshold}

\item{tree}{SNP-scaled phylogenetic tree or list of trees with pairs of tips labelled with sampleA and sampleB. If this is supplied a different model for related SNP distances will be fit that takes into account branch length differences to the MRCA for any given pair of samples provided in sampleA and sampleB.}

\item{sampleA}{tip labels for sampleA; must be in the correct order with respect to sampleB such that the time distance is calculated as SampleA date-SampleB date (even if this allows for negative numbers)}

\item{sampleB}{tip labels for sampleB; see above}

\item{branch_offset}{overide to branch offset to each skellam parameter for divergence correction model}

\item{lambda_bounds}{bounds of rate estimation in SNPs/year/site if given time and site data}

\item{k_bounds}{bounds of related proportion estimation}

\item{intercept_bounds}{bounds of intercept estimation}
}
\value{
a plot of SNP distance over time using ggplot
}
\description{
Creates a plot of snp distance over time
}
