% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_over_time.R
\name{snp_over_time}
\alias{snp_over_time}
\title{SNP dist/time plot}
\usage{
snp_over_time(
  SNPs,
  Time,
  lambda,
  sites = NA,
  snp_threshold,
  title = "SNPs over Time"
)
}
\arguments{
\item{SNPs}{list of SNP distances from a mixed transmission data set}

\item{Time}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{lambda}{mutation rate to plot in the graph (calculated with mixture_snp_cutoff)}

\item{sites}{list of sites considered for each SNP distance in mixed data set; adjusts lambda to be in SNPs/day if provided}

\item{snp_threshold}{a threshold to apply to the mixed data set for considering related data (calculated with mixture_snp_cutoff)}
}
\value{
a plot of SNP distance over time using ggplot
}
\description{
Creates a plot of snp distance over time
}
