% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_over_time.R
\name{snp_over_time}
\alias{snp_over_time}
\title{SNP dist/time plot}
\usage{
snp_over_time(
  mixed_snp_dist,
  unrelated_snp_dist,
  mixed_time_dist,
  mixed_sites = NA,
  truncation_point = 2000,
  max_time = NA,
  title = "SNP Distance Over Time",
  jitter = TRUE,
  p_value = NA,
  ci_data = NA,
  time_limits = c(0, NA),
  under_threshold = FALSE,
  tree = NA,
  sampleA = NA,
  sampleB = NA,
  branch_offset = NA,
  lambda_bounds = c(0, 1),
  k_bounds = c(0, 1),
  intercept_bounds = c(-Inf, Inf)
)
}
\arguments{
\item{mixed_snp_dist}{list of SNP distances from a mixed transmission data set}

\item{unrelated_snp_dist}{list of SNP distances from an unrelated data set}

\item{mixed_time_dist}{list of time differences between samples from each SNP distance in the mixed data set (in days)}

\item{mixed_sites}{list of sites considered for each SNP distance in mixed data set}

\item{truncation_point}{a SNP distance limit for the data, if set to NA will estimate as if there is no limit}

\item{max_time}{the time (in days) utilised to calculate SNP thresholds, only applicable when time differences are provided, if not provided will utilise maximum of supplied times}

\item{title}{title for graph passed to ggplot}

\item{jitter}{whether to jitter data (keeps data above 0 SNPs and 0 time)}

\item{p_value}{optional input to display time randomised p value from mxsure_timerand_test}

\item{ci_data}{optional input for previously calculated CI data (mxsure_ci) to display}

\item{time_limits}{x axis limits passed to ggplot}

\item{under_threshold}{whether to only plot points below the calculated SNP threshold}
}
\value{
a plot of SNP distance over time using ggplot
}
\description{
Creates a plot of snp distance over time
}
